Add eglot for handling interactions with language servers
Another attempt to include language servers to my emacs workflow. This time it
is eglot, a minimalistic interface to language servers. For now we disable
counsel-etags mode and work with xref-find-definitions again (which is being
adviced by eglot). This should not break the workflow over tramp (fallback to
normal TAGS file + xref-find-definitions).

We have to add a small hack to eglot to make it use projectile for detecting
project roots and we also don't want eglot to clutter our mode line which is why
add some workaround, removing eglog--managed-mode from mode-line-misc-info and
adding eta to minor-mode-alist.